+++ R/rxode-options.R
@@ -5,7 +5,34 @@
   }
 }
 .hasUnits <- FALSE
+.rcppVersion <- packageVersion("Rcpp")
+.PreciseSumsVersion <- packageVersion("PreciseSums")
+.dparserVersion <- packageVersion("dparser")
+.checkmateVersion <- packageVersion("checkmate")
 .onLoad <- function(libname, pkgname) { ## nocov start
+  if (!identical(.rcppVersion, packageVersion("Rcpp"))) {
+    stop("RxODE compiled with Rcpp '", as.character(.rcppVersion),
+         "' but Rcpp '", as.character(packageVersion("Rcpp")), "' is loaded\nRecompile RxODE with the new version of Rcpp",
+         call.=FALSE)
+  }
+  if (!identical(.dparserVersion, packageVersion("dparser"))) {
+    stop("RxODE compiled with dparser '", as.character(.dparserVersion),
+         "' but dparser '", as.character(packageVersion("dparser")),
+         "' is loaded\nRecompile RxODE with the this version of dparser",
+         call.=FALSE)
+  }
+  if (!identical(.checkmateVersion, packageVersion("checkmate"))) {
+    stop("RxODE compiled with checkmate '", as.character(.checkmateVersion),
+         "' but checkmate '", as.character(packageVersion("checkmate")),
+         "' is loaded\nRecompile RxODE with the this version of checkmate",
+         call.=FALSE)
+  }
+  if (!identical(.PreciseSumsVersion, packageVersion("PreciseSums"))) {
+    stop("RxODE compiled with PreciseSums '", as.character(.PreciseSumsVersion),
+         "' but PreciseSums '", as.character(packageVersion("PreciseSums")),
+         "' is loaded\nRecompile RxODE with the this version of PreciseSums",
+         call.=FALSE)
+  }
   if (requireNamespace("pillar", quietly = TRUE)) {
     .s3register("pillar::type_sum", "rxEvid")
     .s3register("pillar::type_sum", "rxRateDur")